split out the bash function tests to own file
